DTK-3636 When uninstall is triggered with --name or -d option, .task_id_... file is not saved in correct location PURGE delete fixed

+++ lib/client/operation/service/uninstall.rb
@@ -53,13 +53,15 @@ module DTK::Client
           response = rest_post("#{BaseRoute}/uninstall", post_body)
           path = ClientModuleDir.ret_base_path(:service, service_instance) unless path
           path = path+'/' if path
-          ClientModuleDir.rm_f(path) if args[:purge]
 
+          ClientModuleDir.rm_f(path) if args[:purge]
           if message = response.data(:message) || "DTK service '#{service_instance}' has been uninstalled successfully."
-            Dir.entries(path).each do |f|
-              ClientModuleDir.rm_f(path+f) if f.include? '.task_id_'
+            unless args[:purge] || !Dir.exists?(path)
+              Dir.entries(path).each do |f|
+                ClientModuleDir.rm_f(path+f) if f.include? '.task_id_'
+              end
+              ClientModuleDir.create_file_with_content(path+".task_id_#{response.data(:task_id)}", '') if response.data(:task_id)
             end
-            ClientModuleDir.create_file_with_content(path+".task_id_#{response.data(:task_id)}", '') if response.data(:task_id)
             OsUtil.print_info(message)
           end
         end
